# Conflicts:
#	3rdParty/V8/v8
#	arangod/Transaction/Methods.h
#	arangod/Utils/UserTransaction.h
#	arangod/V8Server/v8-collection.cpp

+++ arangod/Cluster/ClusterEdgeCursor.h
@@ -32,7 +32,7 @@ namespace traverser {
 class ClusterEdgeCursor : public EdgeCursor {
 
  public:
-  ClusterEdgeCursor(arangodb::velocypack::Slice, size_t, ClusterTraverser*);
+  ClusterEdgeCursor(arangodb::velocypack::Slice, uint64_t, ClusterTraverser*);
 
   ~ClusterEdgeCursor() {
   }
